+++ linkup/src/features/meeting/components/FilterDropdown.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { reactive, watch, computed } from 'vue'
-import FilterRegionDropdown from '../../../components/common/FilterRegionDropdown.vue'
+import FilterRegionDropdown from './FilterRegionDropdown.vue'
 
 const props = defineProps({
   filters: Object,
